requirements.txt
facebook-sdk==0.4.0
Flask==0.10.1
Flask-Slack==0.1.2
google-api-python-client==1.3.2
gunicorn==19.2.1
httplib2==0.9
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
oauth2client==1.4.6
pyasn1==0.1.7
pyasn1-modules==0.0.5
rsa==3.1.4
simplejson==3.6.5
six==1.9.0
uritemplate==0.6
Werkzeug==0.10.1
